React Doctor指摘を解消
